# Conflicts:
#	flight/Modules/Actuator/actuator.c
#	flight/Modules/ManualControl/transmitter_control.c
#	flight/Modules/OnScreenDisplay/onscreendisplay.c
#	flight/targets/dtfc/fw/pios_board.c

+++ flight/PiOS/STM32/inc/pios_inlinedelay.h
@@ -46,7 +46,7 @@ static inline void PIOS_INLINEDELAY_AssertClockSpeed()
 	PIOS_Assert(us_ticks = other_us_ticks);
 }
 
-static inline uint32_t PIOS_INLINEDELAY_NsToCycles(uint16_t ns)
+static inline uint32_t PIOS_INLINEDELAY_NsToCycles(uint32_t ns)
 {
 	/* assumptions: integral number of MHz. */
